+++ tests/stubs/views/trans.blade.php
@@ -1,0 +1,20 @@
+{{ trans('JSON trans') }}
+{{ trans_choice('JSON trans_choice', 10) }}
+{{ Lang::get('JSON Lang::get') }}
+{{ Lang::choice('JSON Lang::choice') }}
+{{ Lang::trans('JSON Lang::trans') }}
+{{ Lang::transChoice('JSON Lang::transChoice', 10) }}
+@lang('JSON @lang')
+@choice('JSON @choice', 10)
+{{ __('JSON __') }}
+
+
+{{ trans('php.trans') }}
+{{ trans_choice('php.trans_choice', 10) }}
+{{ Lang::get('php.Lang::get') }}
+{{ Lang::choice('php.Lang::choice') }}
+{{ Lang::trans('php.Lang::trans') }}
+{{ Lang::transChoice('php.Lang::transChoice', 10) }}
+@lang('php.@lang')
+@choice('php.@choice', 10)
+{{ __('php.__') }}