+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to `locale-laravel` will be documented in this file.
 
+## 2.0.0 - 2022-11-22
+
+- New: `--force` option to sync without interaction.
+- New: `--purge` option to delete (on Locale) unused translation.
+- New: Load translation keys from views and application source.
+- New: Update translations for target locales on sync command.
+
+## 1.0.1 - 2022-05-26
+
+- Downgrade composer dependencies to work from Laravel `^7.0|^8.0|^9.0`.
+
 ## 1.0.0 - 2022-05-23
 
 - Initial release